expBar

+++ Assets/01_Scripts/Core/IAgent.cs
@@ -6,7 +6,7 @@ using UnityEngine.Events;
 
 public interface IAgent
 {
-    public int Health { get; }
+    public float Health { get; }
     public UnityEvent OnDie { get; set; }
     public UnityEvent OnGetHit { get; set; }
 }
