+++ Assets/01_Scripts/Core/IHitAble.cs
@@ -6,6 +6,6 @@ public interface IHitAble
 {
     public bool IsEnemy { get; }//메서드도 아니고 변수도 아닌 프로퍼티; 
     public Vector3 HitPoint { get; }
-    public void GetHit(int damage, GameObject damageDealer);
+    public void GetHit(float damage, GameObject damageDealer);
 
 }
